Implement a BUILD_THE_BUILDTOOLS option.
When this variable is set to ON with -DBUILD_THE_BUILDTOOLS=ON, this
puts build_projects in a special mode to build and install only a
small list of buildtools.  Only cmake is on that list of buildtools
for now that are built with this option.  cmake is built with the
bootstrap method to cut the dependence on the binary version of cmake
that is used to configure build_projects in the
-DBUILD_THE_BUILDTOOLS=ON case.

The idea is once the buildtools list of packages is built with this
special option, then you can put those buildtools on the PATH and use
them from then on in normal builds of build_packages to reduce the
reliance on downloaded binaries for all results.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> using the
build_all target for the "Unix Makefiles" generator on Linux and the
"MSYS Makefiles" generator on (the Wine version of) Windows.  Note
these were just successful build tests, and the resulting
cmake-2.8.11.1 binary in each case has not been tested (yet) by using
it for builds of other software.  But all subsequent tests will be
using these bootstrapped versions.


svn path=/trunk/; revision=12403