The overview is the stdout from this script will be used as input for another python script (still being designed) to transform the single complete xml file into build_projects configuration files.
This should be the final version of this script for gtk-3.4.1, but it
is likely changes will have to be made for other versions of gtk
modules (when we move to those) since some exact syntax assumptions
are made that are correct for 3.4.1 jhbuild module files, but might
not be correct for other versions.  

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> by running

./gtk_xml_recursive_process.py /home/software/pango_stack/jhbuild/modulesets/3.4.1/gnome-apps-3.4.1.modules

This command followed all the includes.  That is, all *.modules files
in that directory were included in the output to stdout as measured by
some spot checking and "wc -l" line count results that jibed with the
total line count for *.modules files in that directory (if you account
for three dropped include lines and two head lines and one tail line
that are dropped from the included files).


svn path=/trunk/; revision=12406