Initial commit of script to parse jhbuild xml file produced by gtk_xml_recursive_process.py (which combines included chain of jhbuild xml files into one file).
Currently the principal function of the script, parse_jhbuild only
has the logic tested where if_dependencies is True (i.e., where dependencies
are being recursively followed), and that test on real jhbuild data showed
a circular dependency which lead to infinite depths in the recursion.

So some method of detecting circular dependencies and erroring out
must be implemented, and the logic for the alternative straight
printout of values without following the dependencies must be extended
a bit as well.


svn path=/trunk/; revision=12407