Replace WIN32 AND NOT CYGWIN logic everywhere in CMake logic with the variable WIN32_AND_NOT_CYGWIN.
Force our build system to use the modern usage which does not set
WIN32 on CYGWIN platforms.  Because of this change,
WIN32_AND_NOT_CYGWIN is defined to be the same as ${WIN32}, but keep
the WIN32_AND_NOT_CYGWIN name of the variable just to be clear.

Note the Cygwin platform has no chance to work until a further change
to replace WIN32 everywhere with WIN32_OR_CYGWIN is completed.  That
change should be coming soon.


svn path=/trunk/; revision=12410