Update the initial script so that it
(1) finds all valid direct and indirect dependencies of the
starting_package specified on the command-line by the user according
to a command integer also given on the command line (1 bit ON ==> follow
hard dependencies (jhbuild "dependencies"), 2 bit ON ==> follows soft
dependencies (jhbuild "suggests"), 4 bit ON
==> follow "would be nice" run-time dependencies (jhbuild "after")).

(2) Dump to stdout all relevant build configuration information for the
starting package and all valid dependencies that are found.

This should be the final form of this script or close to it.


svn path=/trunk/; revision=12416