Follow the build of CMake itself so that the initial order of commands is
cmake_minimum_required(...
set(CMAKE_LEGACY_CYGWIN_WIN32 0)
project(...

This reordering should (hopefully) solve a current Cygwin issue where
there is a warning about CMAKE_LEGACY_CYGWIN_WIN32 not being set.

This reordering means that the variable CMAKE_SYSTEM_NAME is not
defined when cmake_minimum_required is called so we must use the same
minimum CMake version for all platforms.  In retrospect this is a good
thing since the value of the minimum version sets a number of CMake
policies, and we want those policies (how CMake logic is interpreted)
to be identical on all platforms.

I also took this opportunity to update the minimum version from 2.8.2
on Linux and 2.8.6 on all other platforms to a uniform 2.8.9.  I chose
that minimum to be consistent with the CMake version available on
Debian Wheezy (released a few months ago, and typically more
conservative about CMake versions than other modern Linux
distributions).  What this means is modern Linux distribution users
(and presumably users of modern Mac versions) should be able to build
PLplot without having to build or download a modern CMake version as
well.  As always, Windows users and legacy Linux distribution users
will have to download a binary CMake or build CMake with
a version that satisifies the version requirements (currently 2.8.9).

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> using the
install target (for -DBUILD_TEST=ON) and also by running ctest.
Note, I view this testing as the minimum acceptable for such a
fundamental build-system change where some of the CMake policies
(interpretations of CMake logic) have been updated on Linux from those
policies in effect for 2.8.2 to the policies in effect for 2.8.9.


svn path=/trunk/; revision=12419