Initial commit of the ffi build configuration.  libffi is a hard dependency of modern glib which is not mentioned in the jhbuild configuration of glib.  So adding this build configuration for libffi is a workaround for that jhbuild configuration bug for glib.
Tested by Alan W. Irwin <airwin@users.sourceforge.net> using the
build_ffi target for both the "Unix Makefiles" generator on Linux and
the "MSYS Makefiles" generator on Wine.  Note the latter was a factor
of 35 (!) slower than the former.  (5 minutes versus 9 seconds.) My
guess is most of this extra time is taken up running the configure
script.  That is, that part of the build is strongly subject to
startup latency costs for Wine so that for builds of much larger
autotooled packages, the Wine slowdown factor will generally be
substantially smaller than 35.


svn path=/trunk/; revision=12433