Implement AUTOTOOLS_PARALLEL_BUILD option which defaults to ON (and which works well on Linux).  It might be useful to set this to OFF for MSYS/Wine builds because that reduces some strain caused by running many large libtool bash scripts simultaneously and which appears to pop some error boxes from time to time under MSYS wine.
Reduce the number of builds that are directly configured since their
dependencies indirectly bring in the rest.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux using
the build_wxwidgets, build_pango, and build_plplot targets.  These tests
are for the recent cumulation of commits up to this one.  wxwidgets
built in parallel with no issues, but because of a minor PATH issue
was ignored by the plplot build.  The pango build (and the build of all its many
dependencies) showed no issues and finished in 7.5 minutes.  The
PLplot build (and all its specific small dependencies) had no issues
(except for ignoring the wxwidgets build results and a few additional
dependency issues to be discussed).  The
cairo device driver was linked with all the pango-related libraries that
were built and the cairo non-interactive devices tested well.  Limited
tests of the cairo non-interactive devices also seemed fine.

In sum, this is a breakthrough for the build_packages project where
for the first time with this project we have built essentially the
latest version of pango and its dependencies without issues, and used
those libraries to produce excellent results for the PLplot cairo
devices on Linux.

ToDo: the libfreetype and libpng dependencies are not yet part of
build_packages so the above PLplot results used the system versions of
those instead on Linux.  There may be other dependencies that are
required (to be discussed on the plplot-devel list) as well.


svn path=/trunk/; revision=12450