Initial commit of a script that makes it convenient to check XML catalogs on MSYS using the xmlcatalog --shell environment.
With two arguments (with the first a reference to a file assumed to be
a catalog file) this script goes into xmlcatalog shell mode to look up
entities with both the public and system attributes.  Otherwise, it
just falls through to the equivalent raw xmlcatalog command.

The only point of this script is it provides a reasonably convenient
human workaround for the raw xmlcatalog issue on MSYS platforms where
that environment mangles entity names on the command line (it
interprets them as a path and does some substitution into the entity
which completely messes it up, see
https://mail.gnome.org/archives/xml/2013-January/msg00009.html).  I
have not been able to figure out how to make this script respond
identically to what the raw xmlcatalog command should actually be
doing if the entities didn't get mangled. So this script is only for
convenient human use to check catalogs, but cannot be used in scripts
itself as a replacement for raw xmlcatalog.


svn path=/trunk/; revision=12464