Sort out all generated catalog issues for the MSYS case.
This gives good results for build_gtk-doc on Linux, but
that build still fails for the MSYS case because of the
entity mangling issue for raw xmlcatalog on that platform
(see https://mail.gnome.org/archives/xml/2013-January/msg00009.html).
However, for the MSYS case have spot-checked a few of the generated
catalog results with the xmlcatalog_wrapper.sh script, and they
seem fine.  Which seems to indicate that the catalogs generated
by the docbook-x?l packages are fine on MSYS.


svn path=/trunk/; revision=12465