Implement new "dblatex --backend=xetex" method for generating pdf form of DocBook documentation by default.  (The older xmlto method is still available using -DDOCBOOK_DBLATEX_BACKEND=OFF).
The (big) advantage of this new method compared to the previous xmlto
method of generating the pdf form of our documentation is it allows a
wide range of UTF-8 glyphs in the DocBook source.  
So if the default fonts selected for the xelatex step in the procedure
are comprehensive (such as FreeSans, FreeSerif, and
FreeMono) a very wide range of UTF-8 glyphs can be rendered properly
in our print documentation.

The disadvantage of this method compared to the previous xmlto
procedure is that PostScript results can only be generated using
pdf2ps.  I have implemented that method but then disabled that part of
the code because the result is literally 50 times larger than the pdf
form of our documentation, takes a very long time to generate, takes a
long time to view, and produces rather ugly results.

ToDo: move from the present rather limited defaults fonts to FreeSans,
FreeSerif, and FreeMono.


svn path=/trunk/; revision=12578