Fix a typo (missing blank delimiter after the fourth quote) in the
string(REGEX REPLACE " " ";" PYQT_SIP_FLAGS "${PYQT_SIP_FLAGS}")

command.  This issue was discovered by CMake-2.8.12.  This should have
produced a wrong number of arguments error in the past for this
command but somehow did not so I think previous CMake versions
did not demand a blank after string to delimit the
next argument so all was well.

Anyhow, this issue did not appear to cause problems in the past for
our pyqt4 results, but nevertheless it is good to fix this issue.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> using
the test_pyqt4_example target.


svn path=/trunk/; revision=12586