Implement PREFER_DOWNLOAD approach by default.  The result for PLplot is (currently) to use a 5.9.10 download plus a patch called plplot.patch to add the important PLplot changes since the 5.9.10 release.  That patch was created with the following command:
diff -Naur --exclude=doc --exclude=plplot_octave_txt \
--exclude=build_projects --exclude=.svn --exclude="*README*" \
--exclude="#*#" --exclude=debian --exclude="*.pyc" --exclude=sys \
--exclude=www --exclude=docbook.cmake $prefix1 $prefix2 \
>| plplot_allura/cmake/build_projects/plplot/plplot.patch

where the excludes get rid of unimportant differences (from the point
of view of building software) between the two prefixes such as changes
in the documentation that have occurred since the 5.9.10 release.
Note, that $prefix1 is the directory location of the top of the source
tree from the unpacked 5.9.10 tarball, and $prefix2 is the top of the
local copy of the source tree created from svn trunk with local
changes potentially added (and also potentially added to the patch).

As a result, the -DPREFER_DOWNLOAD=ON approach that is used by default
for the PLplot build has maintenance issues because it requires
keeping that patch up to date.  On the other hand, it also removes
some uncertainty due to svn head for PLplot being a moving target.
Also, on the other hand, it avoids the considerable downside of using
the subversion approach on Windows which requires (a) a subversion
client be installed which adds the normal security concerns about
running any downloaded Windows binary from some new location on the
web, and (b) all dependencies of that client be installed.  The latter
is especially important for Wine since Windows subversion clients use
ntml authorization.  Wine can only provide that if the Linux platform
underneath Wine has Samba installed which might open a channel for
viruses to propagate from Wine to the underlying Linux platform.  So
my view is -DPREFER_DOWNLOAD=ON has sufficiently fewer security and
moving target downsides that it justifies the maintenance cost of
keeping the patch up to date.  At the same time, if the user has
subversion installed, and they want to follow the moving target of the
svn trunk version of PLplot, they are always free to specify
-DPREFER_DOWNLOAD=OFF.


svn path=/trunk/; revision=12599