Replace PyArray_FLOAT type (which is not part of the clean numpy-1.7 API) with appropriate NPY_* style types.
This fixed the numpy-1.7.1 compile errors caused by

#define NPY_NO_DEPRECATED_API    NPY_1_7_API_VERSION

In two cases (both when calling a single-precision version of PLplot from python)
the PyArray_FLOAT replacement is by NPY_FLOAT32.  That replacement
seemed appropriate but has not been tested.

In the two other cases (when PL_HAVE_PTHREAD is not #defined)
PyArray_FLOAT was replaced by NPY_PLFLT since I am pretty sure that
whether that macro is #defined or not should have nothing to do with
precision.  This case (involving mapform) was tested on the
MinGW/MYSYS/Wine platform, and the results for example 19 were
substantially improved from major differences to just a few small
rounding troubles.  So this change was definitely a bug fix
for the mapform case.

The remaining changes were to use a specific floating-point
precision types in all cases for clarity.



svn path=/trunk/; revision=12606