For the ps, psttf, and hpgl device drivers change all macros concerning pen width to floating point so that no casts are implicitly made when comparing to plsc->width.
Of course, these device drivers all write integer pen widths to the
file so make the appropriate casts concerning that.

ToDo (some day for somebody else).  Investigate whether the
PostScript standard allows floating values for the pen width,
and if so, do not do the cast and instead write out to the floating
point value of the width to the file.


svn path=/trunk/; revision=12612