Convert from 2.1.0 release to 2.2.1 release.
Convert to using patches to express changes to the 2.2.1 release
rather than files which are difficult to adjust for any upstream
libharu changes in future.

Tested by Alan W. Irwin <airwin.users.sourceforge.net> following
directions in README.libharu for a standalone build on Linux
and also using the build_projects approach on Linux.

ToDo: MinGW/MSYS/Wine currently fails (presumably due to
2.1.0 ==> 2.2.1 change).  So will have to figure out
the problem, and update the patches for that case.


svn path=/trunk/; revision=12615