Continue to rationalize and strengthen Itcl, Itk, and Iwidgets detection and consistency checking.
Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux using
the output from cmake configuration of PLplot.  The cases tested were as
follows:

1. Ordinary PATH variable, use (the default)
-DUSE_INCRTCL_VERSION_4=OFF.

2. Identical to (1.), except specify -DUSE_INCRTCL_VERSION_4=ON.

3. Identical to (1.), except the PATH was modified to put a special install location used by
epa_build (a.k.a., build_projects) on the PATH.  The result was the 
tclsh version built by epa_build was used for this test.

4. Identical to (3.), except specify -DUSE_INCRTCL_VERSION_4=ON.

The result for (1.) was consistent data for Tcl8.5, Tk8.5, version 3
of Itcl and Itk, and version 4.0.1 of Iwidgets were found
corresponding to the (old) Debian wheezy packages that have been
installed on my system.  The result for (2.) smoothly dropped Itcl,
Itk, and Iwidgets (as expected since Debian wheezy does not give
access to version 4 of Itcl and friends).  The result for (3.) was
consistent data for Tcl8.6, Tk8.6, version 3 of Itcl and Itk (both
built against Tcl/Tk8.6), and version 4.0.1 of Iwidgets were found
corresponding to the epa_build for version 3 of itcl and friends. The
result for (4.) was consistent data for Tcl8.6, Tk8.6, version 4 of
Itcl and Itk (both built against Tcl/Tk8.6), and version 4.1.0 of
Iwidgets were found corresponding to the epa_build for version 3 of
itcl and friends.

So a simple manipulation of the PATH and/or the
-DUSE_INCRTCL_VERSION_4 cmake option gives consistent results
in all cases.


svn path=/trunk/; revision=12672