Correct possible USE_TCL_STUBS and USE_TK_STUBS error which consist of calling Tcl_InitStubs, Tk_InitStubs, and (redundantly) Tcl_InitStubs (first thing done inside the call to PlbasicInit).  That redundant call might be an issue. Anyhow, eliminate it by calling PlbasicInit (which includes no Tk calls), then Tk_InitStubs.  This is the method used by bindings/tk/Pltk_Init.c.
N.B. We have never tested the Tcl/Tk stubs system because the
build system does not set USE_TCL_STUBS or USE_TK_STUBS and
does not link to the stubs versions of the Tcl/Tk libraries.
So the possible error above has never been tested and the
fix also won't be tested unless and until we "stubify" our
PLplot extensions to Tcl/Tk.  Nevertheless, the Pltk_Init.c
stubs initiation is obviously correct so it is good to use
exactly the same method for Plplotter_Init.c.


svn path=/trunk/; revision=12688