For historical reasons ZEROW2B, ZEROW2D, ONEW2B, and ONEW2D were dropped from the Tcl constants, but since those are correctly #defined in plplot.h and not dropped by any other language include those constants.
Make important fix for all tcl constants in the PLPLOT namespace which
is to process them with expr to convert the hexadecimal form to
decimal.  It turns out the hexadecimal form never worked previously,
but we didn't know that because only constants that were unused in our
tcl standard examples were expressed as hexadecimal.  But since the
recent repropagation effort based on the numerical macros in plplot.h
(many of which were in hexadecimal form) a much larger fraction of the
tcl constants were in hexadecimal form and this (until the present fix
to convert all tcl constants to decimal form) completely screwed up
standard example 33 results for tcl.

Tested by: Alan W. Irwin <airwin@users.sourceforge.net> on Linux
using the test_diff_psc target.


svn path=/trunk/; revision=12692