Implement test_wish_standard_examples target that executes all the standard examples using wish and PLplot's TEA capabilities. Make this target a dependency of test_interactive.
N.B. For some strange reason the display of the plots is currently
getting lost when this script is run non-interactively.  That is, if
you run the script directly or by mentioning it as an argument for the
wish command or if it sourced from an interactive wish environment.
However, if you run the commands in this script interactively, i.e.
cut and paste the script lines directly into a wish environment, the
display of the plots is fine!  See the current question on the
plplot-devel list about how this issue can be fixed.  Nevertheless,
despite this issue the current test does work in the sense that all
the examples run without any error condition.


svn path=/trunk/; revision=12702