Make Tcl logic much more robust for discovering whether the path to the relevant pkgIndex.tcl file is a subdirectory of the build tree.
This fixes an issue with the old logic which was based on whether
the current directory was a subdirectory of the build tree.  That old
logic obviously failed for the common case when the build tree for the installed examples
was actually a subdirectory of the core build tree.  The new logic is
independent of the current directory which is a good thing.


svn path=/trunk/; revision=12709