Use the new name of the template file, epa_CMakeLists.txt.in and change the names of the configured files from */bp.cmake to */CMakeLists.txt
Tested by Alan W. Irwin <airwin@users.sourceforge.net> using the
procedure in README.developers.  The result of this reconfiguration of
most of the build configurations was no changes at all which confirms
the procedure in README.developers is now consistent with the new
*/CMakeLists.txt naming scheme.

N.B. this consistency is a good first step but there is more to
do before the planned add_subdirectory technique for epa_build
works.


svn path=/trunk/; revision=12797