Improve filter_rpath function by taking advantage of the well-documented and maintained CMAKE_C_IMPLICIT_LINK_DIRECTORIES, CMAKE_CXX_IMPLICIT_LINK_DIRECTORIES, and CMAKE_FORTRAN_IMPLICIT_LINK_DIRECTORIES variables to clean rpath information of standard directory locations instead of using the undocumented/poorly maintained internal variable CMAKE_PLATFORM_IMPLICIT_LINK_DIRECTORIES for the same task.
The new set of variables is only available after C, CXX, and FORTRAN
have been enabled so move include of shapelib module (which uses
filter_rpath) to after when all the languages are defined.

This fix sorted out an issue with the rpath used for qt_example which
was being contaminated by a system location and thus giving the wrong
location when a non-standard Qt location is used.


svn path=/trunk/; revision=12848