Update epa_build configuration so that itk (version 4.0.0) successfully builds on MinGW/MSYS.
N.B. A patch corresponding to running "autoreconf -i" on Linux has
been applied rather than just running "autoreconf -i" on all platforms
because autoreconf is broken on MSYS. Note, this patch is not really
necessary for itk because the resulting changes to the configure
script are so small they don't make a practical difference.  However,
I wanted to pioneer applying "autoreconf -i" this way since
configure.in for itk4 is not that different than for itk3, and the
equivalent patch for the itk3 case might make such a large change in
the itk3 configure script that it solves the internal error for that
script that is current stopping the itk3 epa_build.


svn path=/trunk/; revision=12925