Define and use ENABLE_tkX and ENABLE_itkX to help distingush cases where X is present or not along with Tk and Itk.
Define and use generic_interactive_device, the device that
is to be used with generic interactive tests.  Normally, this
device is xwin on Unix (with X) and wingcc on Windows (without X), but
for the test case where X has not been found on Unix, it is ntk.

Tested by Alan W. Irwin <airwin@users.sourceforge.net> on Linux where
the find for X11 was temporarily commented out so nothing X related
was defined, and the build-system logic normally used for the Windows
case could be tested (but with ntk as the generic interactive device).

ldd -r and nm --undefined-only showed no issues for -dev ntk, the
plplottcltk library, and the pltcl executable.  The
test_pltcl_standard_examples and test_tclsh_standard_examples targets
completed without issues (after the ntk cmd array overflow bug found
by example 27 was fixed).  Note all these tests were in the build tree
for the shared library/dynamic devices case so additional more
comprehensive testing of these widespread changes with
scripts/comprehensive_test.sh is recommended in the long term.


svn path=/trunk/; revision=12934