Commit patch to update swig octave support to version 3.8.0.
This patch was implemented by the swig developer, Karl Wette, for the
pending release of swig-3.0.0.  It was obtained using

wget https://github.com/swig/swig/commit/5b167cc12daf9ea275c17fedaefc975450613ab2.patch

and modified by commenting out the first file which is an update to
the swig-3.0.0 changelog. After that change, this swig-3.0.0 patch
does apply cleanly to swig 2.0.11, but has not been tested beyond that
yet for that version of swig.


svn path=/trunk/; revision=12939