Implement ADD_SPECIAL_CONSISTENCY_CHECKING build system option that adds some custom check targets that ordinarily are not added to a normal build because of the extra dependencies.
One of these targets is the already existing
check_plplot_h.inc target (to check consistency
of the octave binding).

Also implement the check_api_xml_consistency target which is an
additional target of this kind.  That target uses Hǎiliàng Wang's
check application (written in Go and available from
https://github.com/hailiang/go-plplot/tree/master/c/check) to check
the consistency of api.xml with plplot.h.


svn path=/trunk/; revision=12961