Adjust to new version of Hǎiliàng Wáng's check programme which now has a -incdir argument to specify where the configured headers included by plplot.h are located.
This change in "check" means it is no longer necessary to copy plplot.h from
the source to the build directory in order to get this
application to work.

Tested on Linux using the check_api_xml_consistency target which
still produces a clean result on the consistency of api.xml with
our API as defined by plplot.h.


svn path=/trunk/; revision=12972