Update Python and Octave documentation strings consistent with doc/docbook/src/api.xml.
This update includes the following interrelated components.

Update api2swigdoc.pl by forming a rename hash to keep track of the
renaming that occurs in bindings/octave/plplot_octave.i, and using the
renamed documentation names for the Octave documentation case.  This
solves a bug that occurred in the previous version where documentation
of the renamed octave functions was lost.

Propagate the improved results from api2swigdoc.pl using the
check_swig_documentation target and copying the comparison file
generated with api2swigdoc.pl to
bindings/swig-support/swig_documentation.i.  This is a major change to
that latter file not only because of the api2swigdoc.pl changes but
also because api.xml has been updated extensively since the last time
the check_swig_documentation target was run.  For example, api.xml
includes UTF-8 characters now so those have been propagated to
bindings/swig-support/swig_documentation.i.

The UTF-8 characters in swig_documentation.i trigger an Octave bug
where help strings which include UTF-8 characters are displayed in a
truncated fashion, i.e., only displayed up to the first such
character.  Such truncated display is not a showstopper bug, but it is
an annoyance.  We have now made an extremely simple test case not
involving PLplot which shows the display bug, and Andrew is
investigating that test case further.

The UTF-8 characters in swig_documentation.i end up in the
swig-generated file bindings/python/plplotc.py which causes run-time
problems since that file has no encoding string on the first or second
line (which is demanded by Python standards). Swig should give users
control of the encoding of that file, but to work around that swig
bug, I have created the _plplotcmodule_fixed custom target which uses
sed (now required to build the Python binding of PLplot) to append "
It uses coding: utf-8" to the first line of that file.

Changed target depends (just in the build tree since the
fixed plplotc.py is automatically installed with no
further action required in the installed examples tree) from
_plplotcmodule to _plplotcmodule_fixed.

The result has been tested on Linux using the test_noninteractive and
test_interactive targets in the build tree and also (using the
CMake-based build system) for the installed examples tree.


svn path=/trunk/; revision=13035