Fix long-standing memory management issue discovered accidentally by a valgrind run on plserver.
The argv argument to main is an array of argc + 1 pointers.  The first
argc of these point to NULL-terminated strings while the last is a
NULL pointer (according to both Linux and Windows documentation of
main that I have read).

Tk_ParseArgv follows this same standard (i.e., writes a NULL
pointer as the argc element of the argv array) when it modifies argc and argv.

Previously there was only room in myargv for myargc = original argc elements, but
that has now been changed to argc+1, and that solves the memory
management issue as expected.

Writing a NULL just off the end of an allocated block of memory in the
heap is generally not a good thing to do and this fix addresses that
issue.  However, plserver has been reliable before this fix so it
appears we were by accident escaping unscathed from this memory
management issue.  So the practical effect of this fix will not be
apparent at the moment, but plserver should be more robust against
further changes which might have generated a segfault without the
present memory management fix.


svn path=/trunk/; revision=13043