Solve scaling issue (https://sourceforge.net/p/plplot/bugs/143/) demonstrated by the extXdrawable_demo application.  Thanks to Jens (https://sourceforge.net/u/jensplplot/profile/) for reporting this issue which I was able to confirm by attempting to resize the extXdrawable_demo GUI.
I have applied the fix suggested by Jens which in retrospect is
obvious; when you change the geometry by resizing the window, the
result is reported back by the call to XGetGeometry, and the scale
factor must be adjusted to compensate for that changed geometry right
after that call. After this fix, resizing the extXdrawable_demo GUI
produced no strange line artifacts (presumably caused by short
overflows) that were present before the fix.


svn path=/trunk/; revision=13044