Straighten out the dependency mess for glib and gt-doc.
gtk-doc has a real dependency on glib so must add that dependency to
gtk-doc.  But that normally would introduce a circular dependency since glib
depends on gtk-doc!  So in addition to the gtk-doc package
change, fix that circular dependency by disabling use of gtk-doc for
glib, and dropping the dependency of glib on gtk-doc.


svn path=/trunk/; revision=13068