Add epa_build configurations for gtk+ and all its dependencies that are not already part of epa_build.  Note to save build time, this is a somewhat "lite" version of gtk+ where the wayland dependency and installed tests have been disabled.
Replace pango by gtk+ in the list of principal packages in the
top-level CMakeLists.txt to suck in all dependencies of gtk+
(including pango and its dependencies) into the epa_build
configuration.

Tested on Linux using the build_gtk+ target.  The resulting build was
a success (no obvious build errors) and required 23 minutes on my
5-year old PC.


svn path=/trunk/; revision=13069