Update build from wxwidgets version 2.8.12 (which does not run correctly on Linux when linked to gtk+ version 3 that is built with epa_build) to to wxwidgets version 3.0.0.
Add configure option --with-gtk=3 to inform wxwidgets-3.0.0 that it is meant
to be built against gtk+ version 3.

Work around build bug introduced as of wxwidgets-3.0.0 for a
separate build tree by forcing an in-source build.

The result builds without issues on Linux and mostly produces
good run-time results with PLplot (see further discussion on list).


svn path=/trunk/; revision=13079