Disable the octave binding for both plplot and plplot_lite.  There is no epa_build of octave yet.  Furthermore, non-epa_built versions of octave will have many dependencies that are likely to clash with the versions of the other PLplot dependencies that are epa_built.
For the plplot (as opposed to plplot_lite) case on the Windows platform
exclude the pango and qt4_lite dependencies and disable the
corresponding PLplot capabilities because pango, qt4_lite, and all
their dependent packages do not epa_build correctly yet on Windows.

Thus, on Windows the only difference between plplot and plplot_lite
at the moment is the former has a dependency on the wxwidgets epa_build
and allows the corresponding wxwidgets device driver to be built
and tested.

On Unix, the differences between plplot and plplot_lite are larger;
the former has a dependency on wxwidgets, pango, and qt4_lite
and allows the corresponding PLplot capabilities to be built
and tested.


svn path=/trunk/; revision=13083