Fix up several comprehensive testing issues with epa_build.
The wxwidgets-related tests are still individually implemented for all
platforms and for the Windows case those tests are dependencies of the
more comprehensive interactive test targets.  However, for the Unix
case those dependencies have been dropped because of "forced mousing"
issue that occurs with gtk+3 variant of wxwidgets that is beginning to
be available (e.g., for epa_build) on Unix.

Implement DROP_GTK_PLUS_2_BUILDS option (which defaults to OFF).  When
that option is turned ON, all targets which are related to gtk+2 only
are dropped.  This avoids build-time errors (for the static drivers
case) or run-time errors (for the dynamic drivers case) due to
clashes with gtk+3 libraries that are potentially used for other
parts of PLplot (e.g., the cairo and wxwidgets device drivers).

Tested on Linux using the build_plplot target of epa_build that is
configured with -DCOMPREHENSIVE_PLPLOT_TEST_INTERACTIVE=ON and
-DCOMPREHENSIVE_PLPLOT_TEST_NONINTERACTIVE=ON

svn path=/trunk/; revision=13088