Configure and test epa_build of qt5_lite.
Tested on Linux by running the build_qt5_lite target for epa_build.
That build of Qt5 took almost an hour (56 min) compared to a similar
build of Qt4 which took roughly one fourth that time.  The issue
appears to be that some of the configure options that allow dropping
components of Qt4 for the epa_build of qt4_lite do not appear to be
available (as yet) for Qt5.  There may be an opportunity
to drop additional parts of the Qt5 build in the future because
I missed a relevant configure option this time or because later
versions of Qt5 may allow dropping more components of the build.
But for now an hour build is acceptable.

Build-time tests were done with PLplot configured with the -DPLPLOT_USE_QT5=ON
option with qmake from the qt5_lite epa_build on the PATH.  That
PLplot build had no showstopper issues, and running ldd -r on drivers/qt.so,
bindings/qt_gui/libplplotqt.so, and  examples/c++/qt_example showed no
linking issues.

Run-time tests were done with PLplot configured and built as above
using the test_all_qt target without any showstopper errors for the
shared library case in the build tree confirming Andrew's similar good
run-time test results with a system version (as opposed to epa_built
version) of Qt5.

One major rendering issue is all text appears to be vertically offset
by half a character height compared to Qt4 results.  Fixing that issue
is next on my agenda.


svn path=/trunk/; revision=13089