Apply empirically derived (using examples/python/test_circle.py) and examples/python/x02) vertical offset for all text for the Qt5 case. This vertical offset is substantial (greater than half the character box height) and roughly constant for most glyphs/fonts.  However, the final value was optimized for the light diagonal cross glyph in test_circle.py. That example shows that other glyphs (except for the well-known asterisk case which is vertically misaligned by font design) have a dispersion in vertical alignments that is larger than for cairo which in turn is larger than for qt with Qt4. Both those latter dispersion results were derived with no empirical vertical offset at all.
An additional issue which confuses the issue concerning the best
vertical alignment for qt with Qt5, cairo, and qt with Qt4 is the font
selection for all those cases is different (as evidenced by different
sizes/shapes of glyphs in all three cases and missing glyphs (i.e.,
Korean for example 24) for the Qt5 case which are present in the other
cases. I assume all these issues are due to Qt5 version 5.2.1 font
selection and font alignment issues which will be addressed for future
Qt5 versions.  And when those Qt5 fixes occur, this empirical vertical
offset will have to be changed or else withdrawn (set to zero).


svn path=/trunk/; revision=13091