As a trial possibility we use a binary downloaded version of gtk+ on Windows.
Tested for the "MinGW Makefiles" generator on Wine-1.6.1.  cmake
apparently configured the cairo devices without issues so that
is a step forward for Walt Brainerd where this was a showstopper.
However, at least on Wine, the gdi.dll
library causes issues at "make install" time with the message

c:\windows\system32\gdi32.dll: file not recognized: File format not
recognized

So this trial is a work in progress.


svn path=/trunk/; revision=13116